Hover Effect on Splash Page & More Multer Progress
