+++ views/profile.handlebars
@@ -1,13 +1,7 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Profile Page</title>
     <link rel="stylesheet" href="/css/profile-page.css">
     <!-- Bootstrap CSS -->
     <link href="https://stackpath.bootstrapcdn.com/bootstrap/5.3.0/css/bootstrap.min.css" rel="stylesheet">
-</head>
+
 <body>
     <div class="container mt-5">
         <div class="row">
@@ -63,7 +57,7 @@
                 <div class="card mb-3" style="max-width: 540px; margin: 20px auto;">
                     <div class="card-body">
                         <h5 class="card-title">Create a New Post</h5>
-                        <form action="/your-upload-route" method="POST" enctype="multipart/form-data">
+                        <form id="new-posts-form">
                             <div class="mb-3">
                                 <label for="post-text" class="form-label">Post Content</label>
                                 <textarea class="form-control" id="post-text" name="post-text" rows="3" required></textarea>
